+++ src/auth/operations/forgotPassword.ts
@@ -34,6 +34,7 @@ async function forgotPassword(incomingArgs: Arguments): Promise<string | null> {
     args = (await hook({
       args,
       operation: 'forgotPassword',
+      context: args.req.context,
     })) || args;
   }, Promise.resolve());
 
@@ -84,9 +85,11 @@ async function forgotPassword(incomingArgs: Arguments): Promise<string | null> {
   const userJSON = user.toJSON({ virtuals: true });
 
   if (!disableEmail) {
+    const serverURL = (config.serverURL !== null && config.serverURL !== '') ? config.serverURL : `${req.protocol}://${req.get('host')}`;
+
     let html = `${t('authentication:youAreReceivingResetPassword')}
-    <a href="${config.serverURL}${config.routes.admin}/reset/${token}">
-     ${config.serverURL}${config.routes.admin}/reset/${token}
+    <a href="${serverURL}${config.routes.admin}/reset/${token}">
+     ${serverURL}${config.routes.admin}/reset/${token}
     </a>
     ${t('authentication:youDidNotRequestPassword')}`;
 
@@ -122,7 +125,7 @@ async function forgotPassword(incomingArgs: Arguments): Promise<string | null> {
 
   await collectionConfig.hooks.afterForgotPassword.reduce(async (priorHook, hook) => {
     await priorHook;
-    await hook({ args });
+    await hook({ args, context: req.context });
   }, Promise.resolve());
 
   return token;