+++ src/auth/operations/login.ts
@@ -48,6 +48,7 @@ async function login<TSlug extends keyof GeneratedTypes['collections']>(
     args = (await hook({
       args,
       operation: 'login',
+      context: args.req.context,
     })) || args;
   }, Promise.resolve());
 
@@ -133,6 +134,7 @@ async function login<TSlug extends keyof GeneratedTypes['collections']>(
     user = (await hook({
       user,
       req: args.req,
+      context: req.context,
     })) || user;
   }, Promise.resolve());
 
@@ -172,6 +174,7 @@ async function login<TSlug extends keyof GeneratedTypes['collections']>(
       user,
       req: args.req,
       token,
+      context: req.context,
     }) || user;
   }, Promise.resolve());
 
@@ -186,6 +189,7 @@ async function login<TSlug extends keyof GeneratedTypes['collections']>(
     overrideAccess,
     req,
     showHiddenFields,
+    context: req.context,
   });
 
   // /////////////////////////////////////
@@ -198,6 +202,7 @@ async function login<TSlug extends keyof GeneratedTypes['collections']>(
     user = await hook({
       req,
       doc: user,
+      context: req.context,
     }) || user;
   }, Promise.resolve());