+++ src/auth/operations/logout.ts
@@ -46,6 +46,7 @@ async function logout(incomingArgs: Arguments): Promise<string> {
     args = (await hook({
       req,
       res,
+      context: req.context,
     })) || args;
   }, Promise.resolve());