+++ src/globals/operations/restoreVersion.ts
@@ -106,6 +106,7 @@ async function restoreVersion<T extends TypeWithVersion<T> = any>(args: Argument
     req,
     overrideAccess,
     showHiddenFields,
+    context: req.context,
   });
 
   // /////////////////////////////////////
@@ -132,6 +133,7 @@ async function restoreVersion<T extends TypeWithVersion<T> = any>(args: Argument
     entityConfig: globalConfig,
     operation: 'update',
     req,
+    context: req.context,
   });
 
   // /////////////////////////////////////