+++ test/uploads/config.ts
@@ -1,10 +1,11 @@
 import path from 'path';
-import { buildConfig } from '../buildConfig';
+import { buildConfigWithDefaults } from '../buildConfigWithDefaults';
 import { devUser } from '../credentials';
 import getFileByPath from '../../src/uploads/getFileByPath';
 import removeFiles from '../helpers/removeFiles';
 import { Uploads1 } from './collections/Upload1';
 import Uploads2 from './collections/Upload2';
+import AdminThumbnailCol from './collections/admin-thumbnail';
 
 export const mediaSlug = 'media';
 
@@ -12,9 +13,15 @@ export const relationSlug = 'relation';
 
 export const audioSlug = 'audio';
 
+export const enlargeSlug = 'enlarge';
+
+export const reduceSlug = 'reduce';
+
+export const adminThumbnailSlug = 'admin-thumbnail';
+
 const mockModulePath = path.resolve(__dirname, './mocks/mockFSModule.js');
 
-export default buildConfig({
+export default buildConfigWithDefaults({
   admin: {
     webpack: (config) => ({
       ...config,
@@ -89,7 +96,14 @@ export default buildConfig({
       upload: {
         staticURL: '/media',
         staticDir: './media',
-        mimeTypes: ['image/png', 'image/jpg', 'image/jpeg', 'image/gif', 'image/svg+xml', 'audio/mpeg'],
+        mimeTypes: [
+          'image/png',
+          'image/jpg',
+          'image/jpeg',
+          'image/gif',
+          'image/svg+xml',
+          'audio/mpeg',
+        ],
         resizeOptions: {
           width: 1280,
           height: 720,
@@ -126,6 +140,11 @@ export default buildConfig({
             formatOptions: { format: 'jpg', options: { quality: 90 } },
           },
           {
+            name: 'accidentalSameSize',
+            width: 320,
+            height: 80,
+          },
+          {
             name: 'tablet',
             width: 640,
             height: 480,
@@ -141,6 +160,99 @@ export default buildConfig({
             name: 'icon',
             width: 16,
             height: 16,
+          },
+        ],
+      },
+      fields: [],
+    },
+    {
+      slug: enlargeSlug,
+      upload: {
+        staticURL: '/enlarge',
+        staticDir: './media/enlarge',
+        mimeTypes: [
+          'image/png',
+          'image/jpg',
+          'image/jpeg',
+          'image/gif',
+          'image/svg+xml',
+          'audio/mpeg',
+        ],
+        imageSizes: [
+          {
+            name: 'accidentalSameSize',
+            width: 320,
+            height: 80,
+            withoutEnlargement: false,
+          },
+          {
+            name: 'sameSizeWithNewFormat',
+            width: 320,
+            height: 80,
+            formatOptions: { format: 'jpg', options: { quality: 90 } },
+            withoutEnlargement: false,
+          },
+          {
+            name: 'resizedLarger',
+            width: 640,
+            height: 480,
+            withoutEnlargement: false,
+          },
+          {
+            name: 'resizedSmaller',
+            width: 180,
+            height: 50,
+          },
+          {
+            name: 'widthLowerHeightLarger',
+            width: 300,
+            height: 300,
+            fit: 'contain',
+          },
+        ],
+      },
+      fields: [],
+    },
+    {
+      slug: reduceSlug,
+      upload: {
+        staticURL: '/reduce',
+        staticDir: './media/reduce',
+        mimeTypes: [
+          'image/png',
+          'image/jpg',
+          'image/jpeg',
+          'image/gif',
+          'image/svg+xml',
+          'audio/mpeg',
+        ],
+        imageSizes: [
+          {
+            name: 'accidentalSameSize',
+            width: 320,
+            height: 80,
+            withoutEnlargement: false,
+          },
+          {
+            name: 'sameSizeWithNewFormat',
+            width: 320,
+            height: 80,
+            formatOptions: { format: 'jpg', options: { quality: 90 } },
+            withoutReduction: true,
+            fit: 'contain',
+          },
+          {
+            name: 'resizedLarger',
+            width: 640,
+            height: 480,
+          },
+          {
+            name: 'resizedSmaller',
+            width: 180,
+            height: 50,
+            // Why fit `contain` should also be set to https://github.com/lovell/sharp/issues/3595
+            withoutReduction: true,
+            fit: 'contain',
           },
         ],
       },
@@ -198,10 +310,11 @@ export default buildConfig({
     },
     Uploads1,
     Uploads2,
+    AdminThumbnailCol,
   ],
   onInit: async (payload) => {
     const uploadsDir = path.resolve(__dirname, './media');
-    removeFiles(uploadsDir);
+    removeFiles(path.normalize(uploadsDir));
 
     await payload.create({
       collection: 'users',
@@ -242,6 +355,25 @@ export default buildConfig({
       collection: audioSlug,
       data: {
         audio: file.id,
+      },
+    });
+
+    // Create admin thumbnail media
+    await payload.create({
+      collection: AdminThumbnailCol.slug,
+      data: {},
+      file: {
+        ...audioFile,
+        name: 'audio-thumbnail.mp3', // Override to avoid conflicts
+      },
+    });
+
+    await payload.create({
+      collection: AdminThumbnailCol.slug,
+      data: {},
+      file: {
+        ...imageFile,
+        name: `thumb-${imageFile.name}`,
       },
     });
   },